feat:Settings/Security password 변경 page 작업 완료

+++ src/main/java/sky/Sss/global/openapi/NkeyOpenApiController.java
@@ -18,6 +18,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import sky.Sss.domain.user.annotation.UserAuthorize;
 import sky.Sss.domain.user.model.LoginSuccess;
 import sky.Sss.global.file.utili.FileStore;
 import sky.Sss.domain.user.model.Status;
@@ -29,7 +30,7 @@ import sky.Sss.global.openapi.service.ApiExamCaptchaNkeyService;
 
 
 @Slf4j
-@RequestMapping("/Nkey/open")
+@RequestMapping("/nkey/open")
 @RestController
 @RequiredArgsConstructor
 public class NkeyOpenApiController {
@@ -40,13 +41,13 @@ public class NkeyOpenApiController {
     private final FileStore fileStore;
 
     @GetMapping("/{userId}")
-    public ResponseEntity loginFailCheck(@PathVariable String userId) throws InterruptedException {
+    public ResponseEntity<?> loginFailCheck(@PathVariable String userId){
         // 실패 횟수 가져온 다음
         Long loginLogCount = userLoginLogService.getCount(userId, LoginSuccess.FAIL, Status.ON);
 
         // 실패횟수가 5번 이하면 200
         if (loginLogCount < 5) {
-            return new ResponseEntity(new Result<>(userId), HttpStatus.OK);
+            return new ResponseEntity<>(new Result<>(userId), HttpStatus.OK);
         }
 
         Map mapKey = apiExamCaptchaNkeyService.getApiExamCaptchaNkey();
@@ -57,38 +58,52 @@ public class NkeyOpenApiController {
             .captchaKey(key)
             .imageName(image).build();
 
-        return new ResponseEntity(new Result<>(captcha), HttpStatus.BAD_REQUEST);
+        return new ResponseEntity<>(new Result<>(captcha), HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping("/image/{fileName}")
-    public Resource getImage(@PathVariable String fileName) throws MalformedURLException {
+    public Resource getImage(@PathVariable String fileName){
         return fileStore.getCaptchaUrlResource(fileName);
     }
 
 
+//    @UserAuthorize
+    @GetMapping("/captcha")
+    public ResponseEntity<CaptchaNkeyDto> getCaptchaNkeyDto() {
+        log.info(" getCaptchaNkeyDto api");
+        return getCaptchaNkeyDtoResponseEntity();
+    }
+
     @GetMapping("/again")
-    public ResponseEntity captchaKEyAgain(@Validated @ModelAttribute CaptchaNkeyDto captchaNkeyDto,
+    public ResponseEntity<?> captchaKEyAgain(@Validated @ModelAttribute CaptchaNkeyDto captchaNkeyDto,
         BindingResult bindingResult
         , HttpServletRequest request) {
 
         if (bindingResult.hasErrors()) {
             return Result.getErrorResult(new ErrorResultDto(bindingResult, ms, request.getLocale()));
         }
-        String fileName = captchaNkeyDto.getImageName();
+
         //받은 이미지 삭제
         try {
-            apiExamCaptchaNkeyService.deleteImage(fileName);
+            String fileName = captchaNkeyDto.getImageName();
+            if (fileName != null) {
+                apiExamCaptchaNkeyService.deleteImage(fileName);
+            }
         } catch (NoSuchFileException e) {
             log.error("deleteAllBatch image error ", captchaNkeyDto.getImageName());
         }
+        return getCaptchaNkeyDtoResponseEntity();
+    }
+
+    private ResponseEntity<CaptchaNkeyDto> getCaptchaNkeyDtoResponseEntity() {
         String key = (String) apiExamCaptchaNkeyService.getApiExamCaptchaNkey().get("key");
         String image = apiExamCaptchaNkeyService.getApiExamCaptchaImage(key);
-
         CaptchaNkeyDto captcha = CaptchaNkeyDto.builder()
             .captchaKey(key)
             .imageName(image).build();
 
-        return new ResponseEntity(captcha, HttpStatus.OK);
+        return new ResponseEntity<>(captcha, HttpStatus.OK);
     }
+
 
 }
